experiment with mdx
