+++ vuu-ui/showcase/src/examples/Table/BASKET/Basket.mdx
@@ -1,0 +1,5 @@
+import { BasketTable } from "./BasketTable";
+
+## Basket example
+
+<BasketTable tableName="basket" />
