+++ vuu-ui/showcase/src/examples/Table/index.ts
@@ -1,6 +1,6 @@
 export * as TableList from "./TableList.examples";
 export * as Table from "./Table.examples";
-export * as BASKET from "./BASKET.examples";
+export * as BASKET from "./BASKET";
 export * as SIMUL from "./SIMUL.examples";
 export * as BigData from "./BigData.examples";
 export * as TableLayout from "./TableLayout.examples";
